chore: modularize swap and add biome lint
