+++ components/motion/swap/constants.ts
@@ -1,0 +1,4 @@
+export const EASE = [0.16, 1, 0.3, 1] as const;
+export const DRAWER_EASE = [0.32, 0.72, 0, 1] as const;
+
+export const SWAP_DRAWER_EASE = DRAWER_EASE;
